split the simulation methods into a new file
