+++ starport/templates/typed/map/simulation.go
@@ -1,0 +1,55 @@
+package maptype
+
+import (
+	"fmt"
+	"path/filepath"
+
+	"github.com/gobuffalo/genny"
+	"github.com/tendermint/starport/starport/pkg/placeholder"
+	"github.com/tendermint/starport/starport/templates/typed"
+)
+
+func moduleSimulationModify(replacer placeholder.Replacer, opts *typed.Options) genny.RunFn {
+	return func(r *genny.Runner) error {
+		path := filepath.Join(opts.AppPath, "x", opts.ModuleName, "module_simulation.go")
+		f, err := r.Disk.Find(path)
+		if err != nil {
+			return err
+		}
+
+		// Create a list of two different indexes to use as sample
+		sampleIndexes := make([]string, 2)
+		for i := 0; i < 2; i++ {
+			for _, index := range opts.Indexes {
+				sampleIndexes[i] += index.GenesisArgs(i)
+			}
+		}
+
+		// simulation genesis state
+		templateGs := `%[2]vList: []types.%[2]v{
+		{
+			%[3]v},
+		{
+			%[4]v},
+	},
+	%[1]v`
+		replacementGs := fmt.Sprintf(
+			templateGs,
+			typed.PlaceholderSimapGenesisState,
+			opts.TypeName.UpperCamel,
+			sampleIndexes[0],
+			sampleIndexes[1],
+		)
+		content := replacer.Replace(f.String(), typed.PlaceholderSimapGenesisState, replacementGs)
+
+		content = typed.ModuleSimulationMsgModify(
+			replacer,
+			content,
+			opts.TypeName,
+			"Create", "Update", "Delete",
+		)
+
+		newFile := genny.NewFileS(path, content)
+		return r.File(newFile)
+	}
+}
