+++ starport/templates/typed/singleton/stargate.go
@@ -518,40 +518,3 @@ cdc.RegisterConcrete(&MsgDelete%[2]v{}, "%[3]v/Delete%[2]v", nil)
 		return r.File(newFile)
 	}
 }
-
-func moduleSimulationModify(replacer placeholder.Replacer, opts *typed.Options) genny.RunFn {
-	return func(r *genny.Runner) error {
-		path := filepath.Join(opts.AppPath, "x", opts.ModuleName, "module_simulation.go")
-		f, err := r.Disk.Find(path)
-		if err != nil {
-			return err
-		}
-
-		// Create a fields
-		sampleFields := ""
-		for _, field := range opts.Fields {
-			sampleFields += field.GenesisArgs(rand.Intn(100) + 1)
-		}
-
-		templateGs := `%[2]v: &types.%[2]v{
-		%[3]v},
-		%[1]v`
-		replacementGs := fmt.Sprintf(
-			templateGs,
-			typed.PlaceholderSimapGenesisState,
-			opts.TypeName.UpperCamel,
-			sampleFields,
-		)
-		content := replacer.Replace(f.String(), typed.PlaceholderSimapGenesisState, replacementGs)
-
-		content = typed.ModuleSimulationMsgModify(
-			replacer,
-			content,
-			opts.TypeName,
-			"Create", "Update", "Delete",
-		)
-
-		newFile := genny.NewFileS(path, content)
-		return r.File(newFile)
-	}
-}
